add Chakra UI v3 support for Storybook
