+++ .storybook/preview.tsx
@@ -1,0 +1,37 @@
+import type { Preview, ReactRenderer } from "@storybook/nextjs-vite";
+import { Provider } from "../app/components/ui/provider"; // TODO: get this working with `@/components/ui/provider`
+import React from "react";
+import { withThemeByClassName } from "@storybook/addon-themes";
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+
+    a11y: {
+      // 'todo' - show a11y violations in the test UI only
+      // 'error' - fail CI on a11y violations
+      // 'off' - skip a11y checks entirely
+      test: "todo",
+    },
+  },
+  decorators: [
+    (Story) => {
+      return (
+        <Provider>
+          <Story />
+        </Provider>
+      );
+    },
+    withThemeByClassName({
+      defaultTheme: "light",
+      themes: { light: "", dark: "dark" },
+    }),
+  ],
+};
+
+export default preview;
