+++ package.json
@@ -32,10 +32,11 @@
     "build-storybook": "storybook build"
   },
   "dependencies": {
-    "@chakra-ui/react": "^3.22.0",
+    "@chakra-ui/react": "^3.29.0",
     "@emotion/react": "^11.14.0",
     "@fontsource/inter": "^5.2.6",
     "@mapbox/search-js-react": "^1.1.0",
+    "@storybook/addon-themes": "^10.0.5",
     "@types/geojson": "^7946.0.16",
     "@types/node": "^20.2.4",
     "@types/react": "^19.1.8",
